+++ java/component/keyple-calypso/src/main/java/org/eclipse/keyple/calypso/transaction/SamResourceManagerFactory.java
@@ -15,16 +15,15 @@ import static org.eclipse.keyple.calypso.command.sam.SamRevision.AUTO;
 import org.eclipse.keyple.core.selection.SeSelection;
 import org.eclipse.keyple.core.selection.SelectionsResult;
 import org.eclipse.keyple.core.seproxy.ReaderPlugin;
+import org.eclipse.keyple.core.seproxy.ReaderPoolPlugin;
 import org.eclipse.keyple.core.seproxy.SeReader;
 import org.eclipse.keyple.core.seproxy.exception.KeypleReaderException;
 import org.eclipse.keyple.core.seproxy.protocol.SeCommonProtocols;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public abstract class AbstractSamResourceManager {
-    private static final Logger logger = LoggerFactory.getLogger(AbstractSamResourceManager.class);
-
-    protected final ReaderPlugin samReaderPlugin;
+public abstract class SamResourceManagerFactory {
+    private static final Logger logger = LoggerFactory.getLogger(SamResourceManagerFactory.class);
 
     public enum AllocationMode {
         BLOCKING, NON_BLOCKING
@@ -32,38 +31,6 @@ public abstract class AbstractSamResourceManager {
 
     /* the maximum time (in tenths of a second) during which the BLOCKING mode will wait */
     protected final static int MAX_BLOCKING_TIME = 1000; // 10 sec
-
-    AbstractSamResourceManager(ReaderPlugin samPlugin) {
-        this.samReaderPlugin = samPlugin;
-    }
-
-    /**
-     * Allocate a SAM resource from the specified SAM group.
-     * <p>
-     * In the case where the allocation mode is BLOCKING, this method will wait until a SAM resource
-     * becomes free and then return the reference to the allocated resource. However, the BLOCKING
-     * mode will wait a maximum time defined in tenths of a second by MAX_BLOCKING_TIME.
-     * <p>
-     * In the case where the allocation mode is NON_BLOCKING and no SAM resource is available, this
-     * method will return null.
-     * <p>
-     * If the samGroup argument is null, the first available SAM resource will be selected and
-     * returned regardless of its group.
-     *
-     * @param allocationMode the blocking/non-blocking mode
-     * @param samIdentifier the targeted SAM identifier
-     * @return a SAM resource
-     * @throws KeypleReaderException if a reader error occurs
-     */
-    abstract public SamResource allocateSamResource(AllocationMode allocationMode,
-            SamIdentifier samIdentifier) throws KeypleReaderException;
-
-    /**
-     * Free a previously allocated SAM resource.
-     *
-     * @param samResource the SAM resource reference to free
-     */
-    abstract public void freeSamResource(SamResource samResource);
 
     /**
      * Create a SAM resource from the provided SAM reader.
@@ -75,7 +42,7 @@ public abstract class AbstractSamResourceManager {
      * @return a {@link SamResource}
      * @throws KeypleReaderException if an reader error occurs while doing the selection
      */
-    protected SamResource createSamResource(SeReader samReader) throws KeypleReaderException {
+    static protected SamResource createSamResource(SeReader samReader) throws KeypleReaderException {
         logger.trace("Create SAM resource from reader NAME = {}", samReader.getName());
 
         samReader.addSeProtocolSetting(SeCommonProtocols.PROTOCOL_ISO7816_3, ".*");
@@ -95,6 +62,33 @@ public abstract class AbstractSamResourceManager {
         return new SamResource(samReader, calypsoSam);
     }
 
+    /**
+     * Instantiate a new SamResourceManager.
+     * <p>
+     * The samReaderPlugin is used to retrieve the available SAM according to the provided filter.
+     * <p>
+     * Setup a plugin observer if the reader plugin is observable.
+     *
+     * @param readerPlugin the plugin through which SAM readers are accessible
+     * @param samReaderFilter the regular expression defining how to identify SAM readers among
+     *        others.
+     * @throws KeypleReaderException throw if an error occurs while getting the readers list.
+     */
+    static public SamResourceManager instantiate(ReaderPlugin readerPlugin, String samReaderFilter) throws KeypleReaderException {
+        return new SamResourceManagerDefault(readerPlugin, samReaderFilter);
+    }
 
+    /**
+     * Instantiate a new SamResourceManager.
+     * <p>
+     * The samReaderPlugin is used to retrieve the available SAM according to the provided filter.
+     * <p>
+     * Setup a plugin observer if the reader plugin is observable.
+     *
+     * @param samReaderPoolPlugin the plugin through which SAM readers are accessible
+     */
+    static public SamResourceManager instantiate(ReaderPoolPlugin samReaderPoolPlugin) {
+        return new SamResourceManagerPool(samReaderPoolPlugin);
+    }
 
 }
